// MyFirstGitHubAnularProject/MyFirstGitHubAnularProject/ClientApp/app/model/dsgwXMLFile.model.ts
export class DSGWXMLFile {
    appDir: string;
    psftpParams: string;
    logStatus: string;
    logFile: string;
    logSizeMB: string;
    logBackup: string;
    logLevel: string;
    logDaily: string;
    threadPriority: string;
    processPriority: string;
    secureMode: string;
    sshParams: string;
    sshPath: string;
    sftpParams: string;
    puttyPath: string;
    deletePath: string;
    retryFileTransfer: string;
}